Deprecate graphql_client_web, improve web example
The web client now lives in `graphql_client::web`, behind the `web`
feature. The crate is still present but deprecated. It now reexports
from graphql_client.
