+++ examples/web/Cargo.toml
@@ -13,7 +13,7 @@ crate-type = ["cdylib"]
 [dependencies]
 graphql_client = { path = "../../graphql_client" }
 graphql_client_web = { path = "../../graphql_client_web" }
-wasm-bindgen = "0.2.12"
+wasm-bindgen = "0.2.43"
 serde = { version = "1.0.67", features = ["derive"] }
 serde_json = "1.0.22"
 lazy_static = "1.0.1"
