--- examples/web/index.js
@@ -1,5 +1,0 @@
-const js = import("./web");
-
-js.then(js => {
-  js.run();
-});
