--- examples/web/package.json
@@ -1,10 +1,0 @@
-{
-  "scripts": {
-    "serve": "webpack-dev-server"
-  },
-  "devDependencies": {
-    "webpack": "^4.28.3",
-    "webpack-cli": "^3.1.2",
-    "webpack-dev-server": "^3.1.14"
-  }
-}
