--- examples/web/webpack.config.js
@@ -1,10 +1,0 @@
-const path = require("path");
-
-module.exports = {
-  entry: "./index.js",
-  output: {
-    path: path.resolve(__dirname, "dist"),
-    filename: "index.js"
-  },
-  mode: "development"
-};
