+++ graphql_client/tests/Germany.graphql
