+++ graphql_client_web/Cargo.toml
@@ -9,28 +9,10 @@ keywords = ["graphql",  "api", "web", "webassembly", "wasm"]
 categories = ["network-programming", "web-programming", "wasm"]
 repository = "https://github.com/graphql-rust/graphql-client"
 
-[lib]
-crate-type = ["cdylib", "rlib"]
-
-[dependencies]
-failure = "0.1.2"
-graphql_client = { path = "../graphql_client", version = "0.7.1" }
-log = "0.4.5"
-serde_json = "1.0.32"
-wasm-bindgen = "0.2.29"
-wasm-bindgen-futures = "0.3.2"
-js-sys = "0.3.5"
-futures = "0.1.25"
-
-[dependencies.web-sys]
-version = "0.3.2"
-features = [
-    "Headers",
-    "Request",
-    "RequestInit",
-    "Response",
-    "Window",
-]
+[dependencies.graphql_client]
+version = "0.7.1"
+path = "../graphql_client"
+features = ["web"]
 
 [dev-dependencies]
 serde = { version = "1", features = ["derive"] }
